chore: linting (#66)
* chore: linting

The constructor of an abstract class can't be used publicly, so don't make it public. This does have the side effect of documenting protected methods, which may make extension and reuse easier.

* chore: remove NetBeans crud

* test: switch order of arguments in assert

The `equals(Object)` method of the class on the left is used in `assert[Not]Equals(...)` but SonarLint wants constants ("the expected result") on the left, so suppress the warning and put the non-constant on the left.